move implementation out of shared module
